pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robust-nids"
version = "0.1.0"
description = "NIDS Application"
requires-python = ">=3.10"

dependencies = [
    # --- Core scientific stack ---
    "numpy>=2.3,<3.0",
    "pandas>=2.3,<3.0",
    "scipy>=1.16,<2.0",
    "scikit-learn>=1.7,<2.0",
    "matplotlib>=3.10,<4.0",
    "seaborn>=0.13,<1.0",

    # --- Machine Learning / DL ---
    "tensorflow>=2.20,<3.0",
    "keras>=3.11,<4.0",
    "xgboost>=3.0,<4.0",
    "lime>=0.2,<1.0",
    "adversarial-robustness-toolbox>=1.20,<2.0",

    # --- Data tools ---
    "altair>=5.5,<6.0",
    "plotly>=6.3,<7.0",
    "pyarrow>=21.0,<22.0",

    # --- Database ---
    "SQLAlchemy>=2.0,<3.0",
    "psycopg[binary]>=3.2,<4.0",

    # --- Web & UI ---
    "streamlit>=1.49,<2.0",
    "streamlit-authenticator>=0.4,<1.0",
    "streamlit-option-menu>=0.4,<1.0",
    "extra-streamlit-components>=0.1,<1.0",

    # --- Utils ---
    "requests>=2.32,<3.0",
    "python-dotenv>=1.0,<2.0",
    "PyJWT>=2.10,<3.0",
    "pyotp>=2.9,<3.0",
    "qrcode>=8.2,<9.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.1,<26.0",
    "pre-commit>=4.3,<5.0",
    "mypy>=1.0,<2.0",
]

[tool.setuptools]
packages = [
    "core",
    "database",
    "ml",
    "services",
    "ui",
    "utils",
    "views"
]

[tool.black]
line-length = 120
